+++ CHANGES.md
@@ -3,3 +3,8 @@ dev-1.6.5.2
 Plan:
 
     1. bugfix
+    2. handle dead time of tasks in delay queue
+
+Fix:
+
+    1. fix loadAll, the task queue should be reentrant